requirements.txt
python-telegram-bot==21.0.1
anthropic==0.40.0
python-dotenv==1.0.0
httpx==0.27.0
apscheduler==3.10.4
yfinance
